Restructure project with src layout and add Docker dev environment
