+++ .github/workflows/pypipublish.yml
@@ -22,5 +22,6 @@ jobs:
         TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
         TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
       run: |
+        # Just build and publish - src layout handles the separation
         python setup.py sdist bdist_wheel
         twine upload dist/*
